+++ packages/babel-plugin-auto-logger/__e2e__/fixtures/custom-options/logger-output-object-custom-props/output.js
@@ -1,29 +1,29 @@
 function sum(a, b) {
   console.log({
-    path: "[custom-options/logger-output-object-custom-props/input.js:1:19]",
-    from: "sum"
+    "path": "[custom-options/logger-output-object-custom-props/input.js:1:19]",
+    "from": "sum"
   });
 
   try {
     const myPromise = new Promise(() => {
       console.log({
-        path: "[custom-options/logger-output-object-custom-props/input.js:4:40]",
-        from: "array-item-0"
+        "path": "[custom-options/logger-output-object-custom-props/input.js:4:40]",
+        "from": "array-item-0"
       });
     });
     myPromise.catch(reason => {
       console.error({
-        path: "[custom-options/logger-output-object-custom-props/input.js:6:32]",
-        from: "memberExpressionCatch",
-        fnArgs: [reason]
+        "path": "[custom-options/logger-output-object-custom-props/input.js:6:32]",
+        "from": "memberExpressionCatch",
+        "fnArgs": [reason]
       });
       const rejectHandler = 'some implementation for sum';
     });
   } catch (ex) {
     console.error({
-      path: "[custom-options/logger-output-object-custom-props/input.js:9:15]",
-      from: "catchClause",
-      fnArgs: [ex]
+      "path": "[custom-options/logger-output-object-custom-props/input.js:9:15]",
+      "from": "catchClause",
+      "fnArgs": [ex]
     });
     const catchBlock = 'demo purpose for sum';
   }
@@ -33,30 +33,30 @@ function sum(a, b) {
 
 function sub(a, b) {
   console.log({
-    path: "[custom-options/logger-output-object-custom-props/input.js:16:19]",
-    from: "sub"
+    "path": "[custom-options/logger-output-object-custom-props/input.js:16:19]",
+    "from": "sub"
   });
 
   try {
     const myPromise = new Promise(() => {
       console.log({
-        path: "[custom-options/logger-output-object-custom-props/input.js:19:40]",
-        from: "array-item-0"
+        "path": "[custom-options/logger-output-object-custom-props/input.js:19:40]",
+        "from": "array-item-0"
       });
     });
     myPromise.catch(reason => {
       console.error({
-        path: "[custom-options/logger-output-object-custom-props/input.js:21:32]",
-        from: "memberExpressionCatch",
-        fnArgs: [reason]
+        "path": "[custom-options/logger-output-object-custom-props/input.js:21:32]",
+        "from": "memberExpressionCatch",
+        "fnArgs": [reason]
       });
       const rejectHandler = 'some implementation for sub';
     });
   } catch (ex) {
     console.error({
-      path: "[custom-options/logger-output-object-custom-props/input.js:24:15]",
-      from: "catchClause",
-      fnArgs: [ex]
+      "path": "[custom-options/logger-output-object-custom-props/input.js:24:15]",
+      "from": "catchClause",
+      "fnArgs": [ex]
     });
     const catchBlock = 'demo purpose for sub';
   }
@@ -66,30 +66,30 @@ function sub(a, b) {
 
 function multiply(a, b) {
   console.log({
-    path: "[custom-options/logger-output-object-custom-props/input.js:31:24]",
-    from: "multiply"
+    "path": "[custom-options/logger-output-object-custom-props/input.js:31:24]",
+    "from": "multiply"
   });
 
   try {
     const myPromise = new Promise(() => {
       console.log({
-        path: "[custom-options/logger-output-object-custom-props/input.js:34:40]",
-        from: "array-item-0"
+        "path": "[custom-options/logger-output-object-custom-props/input.js:34:40]",
+        "from": "array-item-0"
       });
     });
     myPromise.catch(reason => {
       console.error({
-        path: "[custom-options/logger-output-object-custom-props/input.js:36:32]",
-        from: "memberExpressionCatch",
-        fnArgs: [reason]
+        "path": "[custom-options/logger-output-object-custom-props/input.js:36:32]",
+        "from": "memberExpressionCatch",
+        "fnArgs": [reason]
       });
       const rejectHandler = 'some implementation for multiply';
     });
   } catch (ex) {
     console.error({
-      path: "[custom-options/logger-output-object-custom-props/input.js:39:15]",
-      from: "catchClause",
-      fnArgs: [ex]
+      "path": "[custom-options/logger-output-object-custom-props/input.js:39:15]",
+      "from": "catchClause",
+      "fnArgs": [ex]
     });
     const catchBlock = 'demo purpose for multiply';
   }
@@ -99,30 +99,30 @@ function multiply(a, b) {
 
 function division(a, b) {
   console.log({
-    path: "[custom-options/logger-output-object-custom-props/input.js:46:24]",
-    from: "division"
+    "path": "[custom-options/logger-output-object-custom-props/input.js:46:24]",
+    "from": "division"
   });
 
   try {
     const myPromise = new Promise(() => {
       console.log({
-        path: "[custom-options/logger-output-object-custom-props/input.js:49:40]",
-        from: "array-item-0"
+        "path": "[custom-options/logger-output-object-custom-props/input.js:49:40]",
+        "from": "array-item-0"
       });
     });
     myPromise.catch(reason => {
       console.error({
-        path: "[custom-options/logger-output-object-custom-props/input.js:51:32]",
-        from: "memberExpressionCatch",
-        fnArgs: [reason]
+        "path": "[custom-options/logger-output-object-custom-props/input.js:51:32]",
+        "from": "memberExpressionCatch",
+        "fnArgs": [reason]
       });
       const rejectHandler = 'some implementation for division';
     });
   } catch (ex) {
     console.error({
-      path: "[custom-options/logger-output-object-custom-props/input.js:54:15]",
-      from: "catchClause",
-      fnArgs: [ex]
+      "path": "[custom-options/logger-output-object-custom-props/input.js:54:15]",
+      "from": "catchClause",
+      "fnArgs": [ex]
     });
     const catchBlock = 'demo purpose for division';
   }