+++ packages/babel-plugin-auto-logger/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v-next
+* :house: Update dependencies
+* :lollipop: Use stringLiteral instead of identifier when building object keys
 
 ## v1.2.1 - 2020-04-19
 * :lollipop: ability to log arguments as object when `output.type` is `simple` and `argsAsObject`is `true`